requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
